+++ tests/fixest_tests.R
@@ -1564,7 +1564,7 @@ for(id_fun in 1:5){
     for(s in c("setosa", "versicolor", "virginica")){
         for(lhs in c("y1", "y2")){
             for(rhs in c("x2", "x3")){
-                res = estfun(xpd(..lhs ~ x1 + ..rhs, ..lhs = lhs, ..rhs = rhs), base[base$species == s, ], notes = FALSE)
+                res = estfun(.[lhs] ~ x1 + .[rhs], base[base$species == s, ], notes = FALSE)
 
                 test(coef(est_multi[[k]]), coef(res))
                 test(se(est_multi[[k]], cluster = "fe3"), se(res, cluster = "fe3"))
@@ -1598,6 +1598,37 @@ for(id_fun in 1:5){
     cat("|")
 }
 cat("\n")
+
+
+# No error tests
+# We test with IV + possible corner cases
+
+base$left = rnorm(150)
+base$right = rnorm(150)
+
+est_multi = feols(c(y1, y2) ~ sw0(x1) | sw0(species) | x2 ~ x3, base)
+
+# We check a few
+est_a = feols(y1 ~ 1 | x2 ~ x3, base)
+est_b = feols(y1 ~ x1 | species | x2 ~ x3, base)
+est_c = feols(y2 ~ 1 | x2 ~ x3, base)
+
+test(coef(est_multi[lhs = "y1", rhs = "^1", fixef = "1"]), coef(est_a))
+test(coef(est_multi[lhs = "y1", rhs = "x1", fixef = "spe"]), coef(est_b))
+test(coef(est_multi[lhs = "y2", rhs = "^1", fixef = "1"]), coef(est_c))
+
+# with fixed covariates
+est_multi_LR = feols(c(y1, y2) ~ left + sw0(x1*x4) + right | sw0(species) | x2 ~ x3, base)
+
+est_a = feols(y1 ~ left + right | x2 ~ x3, base)
+est_b = feols(y1 ~ left + x1*x4 + right | species | x2 ~ x3, base)
+est_c = feols(y2 ~ left + right | x2 ~ x3, base)
+
+test(coef(est_multi_LR[lhs = "y1", rhs = "!x1", fixef = "1"]), coef(est_a))
+user_name = c("fit_x2", "left", "x1", "x4", "x1:x4", "right")
+test(names(coef(est_multi_LR[lhs = "y1", rhs = "x1", fixef = "spe"])), user_name)
+test(coef(est_multi_LR[lhs = "y1", rhs = "x1", fixef = "spe"]), coef(est_b)[user_name])
+test(coef(est_multi_LR[lhs = "y2", rhs = "!x1", fixef = "1"]), coef(est_c))
 
 
 ####